fix post /rewards tier

+++ source/RewardsUI/rewards/forms.py
@@ -3,7 +3,7 @@ from django import forms
 
 class RewardForm(forms.Form):
     email_address = forms.EmailField(max_length=100)
-    amount = forms.DecimalField(max_digits=5)
+    amount = forms.IntegerField()
 
 
 class CustomerForm(forms.Form):
